refac: entity - shared의 types 이동
