+++ src/entities/comment/index.ts
@@ -1,0 +1,3 @@
+// Comment 엔티티 배럴 export
+export * from "./api"
+export * from "./model/types"
