+++ src/entities/post/index.ts
@@ -1,0 +1,3 @@
+// Post 엔티티 배럴 export
+export * from "./api"
+export * from "./model/types"
